moved db setup to the api.go file
