+++ api.go
@@ -5,6 +5,10 @@ import (
 	"fmt"
 	"log"
 	"net/http"
+	"context"
+	"os"
+	"go.mongodb.org/mongo-driver/v2/mongo"
+	"go.mongodb.org/mongo-driver/v2/mongo/options"
 )
 
 func WriteJSON(w http.ResponseWriter, status int, v any) error {
@@ -39,7 +43,32 @@ func NewAPIServer(listenAddr string) *APIServer {
 	}
 }
 
+func (s *APIServer) connectToDatabase() {
+	var uri string
+	if uri = os.Getenv("MONGO_URI"); uri == "" {
+		log.Fatal("You must set your 'MONGODB_URI' environment variable. See\n\t https://docs.mongodb.com/drivers/go/current/usage-examples/")
+	}
+
+	serverAPI := options.ServerAPI(options.ServerAPIVersion1)
+	opts := options.Client().ApplyURI(uri).SetServerAPIOptions(serverAPI)
+
+	// Create a new client and connect to the server
+	client, err := mongo.Connect(opts)
+
+	if err != nil {
+		panic(err)
+	}
+
+	defer func() {
+		if err = client.Disconnect(context.TODO()); err != nil {
+			panic(err)
+		}
+	}()
+}
+
 func (s *APIServer) Run() {
+	fmt.Println("Pinged your deployment. You successfully connected to MongoDB!")
+
 	router := http.NewServeMux()
 
 	router.HandleFunc("/account", makeHTTPHandleFunc(s.handleAccount))
