+++ main.go
@@ -1,39 +1,9 @@
 
 package main
 
-import (
-	"context"
-	"fmt"
-	"log"
-	"os"
-	"go.mongodb.org/mongo-driver/v2/mongo"
-	"go.mongodb.org/mongo-driver/v2/mongo/options"
-)
-
 func main() {
-	var uri string
-	if uri = os.Getenv("MONGO_URI"); uri == "" {
-		log.Fatal("You must set your 'MONGODB_URI' environment variable. See\n\t https://docs.mongodb.com/drivers/go/current/usage-examples/")
-	}
-
-	serverAPI := options.ServerAPI(options.ServerAPIVersion1)
-	opts := options.Client().ApplyURI(uri).SetServerAPIOptions(serverAPI)
-
-	// Create a new clietn and connect to the server
-	client, err := mongo.Connect(opts)
-
-	if err != nil {
-		panic(err)
-	}
-
-	defer func() {
-		if err = client.Disconnect(context.TODO()); err != nil {
-			panic(err)
-		}
-	}()
-
-	fmt.Println("Pinged your deployment. You successfully connected to MongoDB!")
 
 	server := NewAPIServer(":3000")
+	server.connectToDatabase();
 	server.Run();
 }
